backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
openai-whisper==20231117
transformers==4.36.2
torch==2.1.2
sentence-transformers==2.3.1
langchain==0.1.0
langchain-community==0.0.10
faiss-cpu==1.7.4
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.2.0
pymongo==4.6.1
gtts==2.5.0
pydub==0.25.1
requests==2.31.0
python-dotenv==1.0.0